Implementation of service catalog
This is an initial implementation of the service catalog, consisting
of three parts.

Service Controller

The service controller performs the following functionality:

- watches for and manages the kubernetes resources in the system
- interacts with service brokers
- handles consumption of binding credential information

Kubernetes Broker

The k8s broker is a white-box service broker which supports
user-defined service classes implemented in terms of Helm charts.
This broker consists of two parts: the broker, which implements
the standard Cloud Foundry Service Broker v2 API; and the registry,
which provides an API for dynamic management of classes in the
broker's catalog.

User-provided Service Broker

The UPS broker provides the "user-provided-service" service class,
which takes as inputs a set of credentials and returns those
credentials when being bound to. It is meant to mimic the Cloud
Foundry concept of "user-provided service instances" using the
standard broker/class model.

The authors are:

Ville Aikas (@vaikas-google)
Brendan Melville (@bmelville)
Martin Maly (@martinmaly)
Michael Kibbe (@mkibbe1993)