Add script to initialize system before CI build
The init_build.sh script updates versions of dependencies,
such as go language or Helm, when running on Jenkins build worker.
This allows a commit to upgrade version of dependencies and even
though the Jenkins worker image may have a different version of
the dependency preinstalled, the dependency will be upgraded to
match the version required by the commit being processed.