Add boilerplate verify scripts
This is an alterntaive to #195.

Rather than spending hours discussing vendoring vs subtrees, I decided it
might be easier to allow people to see the differences and compare them
since it only took a few minutes to throw this together.

I don't think this is fundamentally different from #195 except it vendors
in the other repo instead of using a git subtree.

This one should pass CI though.

Signed-off-by: Doug Davis <dug@us.ibm.com>