Add an href checker for our md files
Verify simple checker for the links/hrefs in our md files.
It will check for files in our tree and http links.

I'm sure there's a better way to do it using awk, but I don't know
awk so if someone wants to offer an alternative/better way I'm all for it.

Signed-off-by: Doug Davis <dug@us.ibm.com>