Add support for running just certain unit tests
new env vars for "make":
TEST_DIRS=...   list of dirs to run 'make test' in
UNIT_TESTS=...  regexp of unit test names to run

they can be combined on the same "make test".

Created a dedicated "test-unit" Makefile target in prep for some
e2e tests later

While we tell people to set NO_DOCKER to "1", I made it so that we're
actually more lax on this and allow it to be set to anything.

Signed-off-by: Doug Davis <dug@us.ibm.com>