Enhance unit testing for cred serialization
Similar to the fuzz testing done in serializer_test, this tests the
serialize function at a higher level from the Inject prospective. The
other difference is that serialization is compared in one direction
only, so no unmarshalling is required to verify the result. Ints,
floats, strings, slices, and maps are data types tested to be correct so
far.