using random port numbers for each test
each test shuts down their server, but doing so doesn’t guarantee that
the server has shut down by the time the shutdown function returns.

This is because the shutdown logic closes a channel, which just sends a
signal to start shutdown.

The change herein avoids this problem by assigning each server a random
port between 3000 and 65533, so that each server is likely to get
assigned a different port than other existing servers (even if they are
in the process of shutting down)