Empty storageType should be treated as "etcd" value
The helm chart is expecting a value `storageType`, if it isn't set to anything
at all helm will fail to do the equal comparison. Adding it as an input value
even when it's empty if could do the check between string and empty string and
everything will work as expected.