Fix getFreshApiserverAndClient server shutdown
This fix simply passes the stopCh in the options so that the deployed
server in run_server is monitoring the right channel for a close event.
This change was also tested with the port randomization removed so that
any existing running servers could be easily identified.

A stop channel is generated for callers that don't provide one in the
options, similar to how things were working previously.

Closes #434