Fix getFreshApiserverAndClient server shutdown (#492)
* Fix getFreshApiserverAndClient server shutdown

This fix simply passes the stopCh in the options so that the deployed
server in run_server is monitoring the right channel for a close event.
This change was also tested with the port randomization removed so that
any existing running servers could be easily identified.

A stop channel is generated for callers that don't provide one in the
options, similar to how things were working previously.

Closes #434

* Change newTestController to return server shutdown function

The shutdown function should only be called once a test is complete,
not before.

* Explicitly create stop channel in server.go

This sets up an interrupt handler to close the stop channel upon
receiving a signal to stop. I tested to see if at the time of the
handler running if the stop channel was already closed, it wasn't.