Publishing docker images from master
Images are published on each commit to master. In the publishing step,
docker images will be tagged with the git tag if there is one (as
reported by travis) or otherwise will be tagged with the commit sha.
Thus, images will never be pushed to an existing tag and will be
immutable.

Fixes #504
Fixes #488