Adding INFO logs for when API server validations fail
@vaikas-google and @MHBauer met earlier today to investigate
https://github.com/kubernetes-incubator/service-catalog/issues/541, and
found the following:

- The ups-broker was sending a catalog with UUIDs with upper-case
letters in it
- The API server disallows upper case letters (it uses
`NameIsDNS1035Label` to check UUIDs)
- The API server received the POST for the new service class, and fails
the validation
- The API server sends the failure back to the core API machinery,
which appears to send error details back to the client
- The client’s `Create` function for service classes returns an error
that doesn’t contain any descriptive detail

This PR is a stop-gap measure to increase the amount of error reporting
we have. Ultimately, we’ll have to determine where the error detail is
being lost (my money is on the generated client, fwiw) and fix it.

cc/ @vaikas-google @krancour @pmorie @MHBauer