Fix up Kubeconfig handling
This commit uncomments and fixes up the Kubeconfig handling, making
it possible to actually pass in Kubeconfigs to use to connect to
the k8s API server and the service catalog API server.

Furthermore, we explicitly try and use in-cluster config for service
catalog API server connections, avoiding a generic error message about
it "possibly" not working in favor of our own warning, which advises
viewers to make sure that the service catalog APIs are registered with
the aggregator.