removing dependency installation pre-requisites from build targets
Making .init a pre-requisite of every build target makes common
actions, like ‘test’ and ‘build’, to take an excessive amount of time.

Generally, speaking, the biggest contributor to the high latency is all
the ‘glide install’ commands that run. Even if the vendor directory is
completely up to date, that command runs and scans the entire vendor
directory. As we know, that directory is large and the scan takes a
long time.

I am removing the dependency here in favor of asking users to manually
run ‘make deps’ in order to install dependencies. I also believe that
it is sufficient to ask developers to periodically ‘refresh’ their
vendor directory as we update our dependencies.