Fix gofmt issue
When running "make verify" I see this:
```
Running gofmt:
Running golint and go vet:
pkg/apis/servicecatalog/checksum/checksum_test.go
```

What's odd is that the file name is due to the gofmt not the golint.
But, the real issues are:
1) the `&>` in the Makefile isn't getting picked up
2) the gofmt issue

This PR replaces `&>` with `> .out 2>&1`.
I don't need to fix that file any more because its been removed.

Signed-off-by: Doug Davis <dug@us.ibm.com>