Check if file permissions allow go install
Currently the integration tests install package dependencies regardless
of whether or not GOROOT permissions allow it. This change will check
for write access on the appropriate directory and if write permissions
aren't found, the -i flag is dropped.

Since there could be a scenario where one wants to install the
dependencies and has a non-writable GOROOT, the environment variable
PKGDIR can be defined in that case.