Use $ and console to indicate multi-command blocks
The console code block type helps highlight the command and its
corresponding output by coloring lines starting with $ in black
and the rest (presumably command output) in blue. This is very helpful
when a block contains a mix of commands and their output.

When there is only a single command, and no output, by omitting the
console block type, the command is displayed in black for consistency.

This is optional formatting, but I've found that it helps the user
experience when going through the docs.