Modification to provision parameters
This PR adds a --params flag to the provision operation, enabling the user
to provision with parameters encoded as JSON rather than individual -p or --param
options. This is useful when the parameters cannot be represented as simple key=value
pairs.

This PR additionally fixes two typos in the parameters.go file.

Closes: #1767