Ensure the controller never performs a no-op status update
Prior to this commit, the controller would, in some cases, post a status
update to the API server where the object's status would be exactly the
same as the object's existing status.

The biggest problem with these no-op status updates is that they result
in the "the object has been modified; please apply your changes to the
latest version and try again" error being logged during
integration tests. The error is logged because the test deletes the
object while the controller is performing a no-op status update.