don't overwrite generated files before verifying
.generate_files make target would call the generators to generate before
calling verify. This would write the thing to be verified, and then
verify what it just wrote, vs verify the current state of the files as
they were when verify was initially called.

I don't know if the extra slash is an issue, but it should not be
needed.

comments on client sepcific generator order

generated openapi updates