Wait for caches to sync before starting up the controller (fixes #1907) (#1915)
If the controller starts reconciling service bindings before the
service instance cache is synced, it won't be able to retrieve the
instance from the cache, making it appear as though the instance
does not exist at all. The controller should therefore wait for all
caches to be synced before performing any work.