Ensure ServiceBinding gets deleted when ServiceInstance updated to an invalid plan after a successful bind
The reconcileServiceBindingDelete() method, before it sent the unbind
request to the broker, tried to retrieve the ClusterServicePlan referenced
in the ServiceInstance that the ServiceBinding points to. If that plan
did not exist (such as in cases when the user changed the plan on the
instance), the binding could not be deleted. The controller must be able
to perform an unbind even if the plan doesn't exist, so it shouldn't
retrieve the plan at all.

This commit changes the controller so it no longer retrieves the plan.
Additionally, the controller now gets the plan ID from Instance.Status.
ExternalProperties, which holds the correct plan id needed for unbind
(whereas the plan in Instance.Spec.ClusterServicePlanRef may be no
longer match the plan ID used when the bind request was performed).