Add JSON and YAML output to svcat (#1944)
* Initial WIP commit for adding JSON/YAML output to svcat

This adds JSON/YAML output to svcat, but does not currently include
the Type Metadata, so we lose the Kind/GroupVersion info.

* Adding JSON and YAML output capabilities

Note: This is not 100% complete. Right now, we do not obtain the
metav1.TypeMeta within svcat. We will need to figure out how to
populate this.

See: https://github.com/kubernetes/client-go/issues/308#issuecomment-378165425

Fixes: #1814

* Code review comments

* Code review comments

* Added boilerplate on json.go