Revert the apiserver flag change in the api-serverdeployment template
This reverts the apiserver-deployment.yaml template for the helm chart.
A recent commit rebased everything onto k8s 1.10. In this, there was
a change to replace the old admission control flag with two new ones.
The old flag is deprecated but still works. In the current state, the
chart does not work on older versions of k8s.

This is a tactical fix, we should probably enhance the chart to have
conditional logic on which it uses so that we can continue to support
people using older k8s versions and stay up to date with latest apiserver
changes.

Fixes #1962