controller: update lastTransitionTime when reason changes
Currently, we update lastTransitionTime only when status changes.
However, we should also update the lastTransitionTime when reason
changes because the condition overall is reflected as "changed" to
the user.

More logically, each condition can be viewed in a state diagram.
When the reason changes, we move from one state to another.