Add a nodeSelector to the chart (#2148)
* Add a nodeSelector to the chart

This adds a nodeSelector to the api and controller manager deployments.
With this, you can specify a node selector for the api and controller manager
deployments. Currently, deployments will use the same nodeSelector. Should we separate them?

This is useful in the case of a mixed cluster with linux and windows nodes. Currently you can run
into issues where the two are scheduled on the windows nodes and that doesn't work.

Closes: #1769

* Add readme

* Refactor to split nodeSelector into two