Controller redesign: perform all steps in a single reconciler invocation
Previously, the controller performed only one step per invocation of the
reconcile function. It then updated the object (it's status) and exited.
When the watch event caused by the update came back to the controller,
it continued with the next step. This method of operation does not fit
well with exponential backoff on errors, because a successfully
performed step resets the backoff, rendering the whole backoff system
useless.

This commit solves this problem by performing all necessary actions
in a single invocation of the reconcile function. At the end of each step
(when it performs the status update), the controller takes the updated
object and works with it instead of the original object.

The watch events that come into the controller (as a result of the
status updates) are now ignored. This prevents the controller from doing
the same work multiple times.