Clean up resolveServiceClassRef & similar functions
These functions try to resolve the service class/plan. If they resolve
it successfully, they just modify the instance object, but don't
actually update it on the server. If they can't resolve it,
the do update the object on the server and they also return an error.

It would be cleaner if they only returned an error and let the caller
perform the actual update. This way, the functions' behavior is
consistent between the two possible outcomes.

This commit does exactly that.